+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Material UI navigation
 - Tag filter component
+- Google firebase functions
 
 ## [0.1.0]
 ### Added